functions/requirements.txt
firebase_functions~=0.4.2
fastembed~=0.2.6
google-cloud-firestore~=2.19.0